% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-architecture.R
\name{nma_audit}
\alias{nma_audit}
\title{Retrieve the audit trail}
\usage{
nma_audit(x)
}
\arguments{
\item{x}{An \code{nmaflow_data} object.}
}
\value{
Audit data frame.
}
\description{
Retrieve the audit trail
}
