% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{nma_bayes_summary}
\alias{nma_bayes_summary}
\title{Bayesian relative-effect summary}
\usage{
nma_bayes_summary(fit, all_contrasts = TRUE, predictive = FALSE, ...)
}
\arguments{
\item{fit}{Bayesian \code{nmaflow_fit}.}

\item{all_contrasts}{Return all treatment contrasts.}

\item{predictive}{Include the predictive distribution for random-effects models.}

\item{...}{Additional arguments to \code{multinma::relative_effects()}.}
}
\description{
Bayesian relative-effect summary
}
