% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{nma_boot_coef}
\alias{nma_boot_coef}
\title{Bootstrap coefficient intervals}
\usage{
nma_boot_coef(fit, B = 1999, seed = getOption("nmaFlow.seed", 260819), ...)
}
\arguments{
\item{fit}{A netmeta-backed fit.}

\item{B}{Number of replicates.}

\item{seed}{Seed.}

\item{...}{Passed to \code{\link[=nma_boot]{nma_boot()}}.}
}
\description{
Bootstrap coefficient intervals
}
