% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{nma_boot_diagnose}
\alias{nma_boot_diagnose}
\title{Diagnose bootstrap stability}
\usage{
nma_boot_diagnose(x)
}
\arguments{
\item{x}{An \code{nmaflow_boot} object.}
}
\value{
List with success rate, unique failure messages and Monte Carlo summaries.
}
\description{
Diagnose bootstrap stability
}
