% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{nma_boot_predict}
\alias{nma_boot_predict}
\title{Bootstrap a user-defined prediction statistic}
\usage{
nma_boot_predict(
  fit,
  prediction_function,
  B = 1999,
  seed = getOption("nmaFlow.seed", 260819),
  ...
)
}
\arguments{
\item{fit}{A netmeta-backed fit.}

\item{prediction_function}{Function returning a named numeric prediction vector.}

\item{B}{Number of replicates.}

\item{seed}{Seed.}

\item{...}{Passed to \code{\link[=nma_boot]{nma_boot()}}.}
}
\description{
Bootstrap a user-defined prediction statistic
}
