% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex-evidence.R
\name{nma_component}
\alias{nma_component}
\title{Fit component network meta-analysis}
\usage{
nma_component(fit, inactive = NULL, sep_components = "+", ...)
}
\arguments{
\item{fit}{A standard netmeta-backed \code{nmaflow_fit} where complex treatment names encode
components using a separator.}

\item{inactive}{Optional inactive component.}

\item{sep_components}{Component separator.}

\item{...}{Passed to \code{netmeta::netcomb()}.}
}
\value{
An \code{nmaflow_fit} object.
}
\description{
Fit component network meta-analysis
}
