% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{nma_contribution}
\alias{nma_contribution}
\title{Contribution matrix or weights}
\usage{
nma_contribution(fit, ...)
}
\arguments{
\item{fit}{An \code{nmaflow_fit} object.}

\item{...}{Backend arguments.}
}
\description{
Contribution matrix or weights
}
