% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex-evidence.R
\name{nma_crossdesign}
\alias{nma_crossdesign}
\title{Fit cross-design and cross-format NMA}
\usage{
nma_crossdesign(..., run = TRUE, run_args = list())
}
\arguments{
\item{...}{Arguments passed to \code{crossnma::crossnma.model()}.}

\item{run}{Whether to run JAGS after building the model.}

\item{run_args}{Named list passed to \code{crossnma::crossnma()} when \code{run=TRUE}.}
}
\value{
Model specification or fitted \code{nmaflow_fit}.
}
\description{
Fit cross-design and cross-format NMA
}
