% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nma_design_by_treatment}
\alias{nma_design_by_treatment}
\title{Design-by-treatment inconsistency decomposition}
\usage{
nma_design_by_treatment(fit, ...)
}
\arguments{
\item{fit}{A netmeta-backed fit.}

\item{...}{Backend arguments.}
}
\description{
Design-by-treatment inconsistency decomposition
}
