% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{nma_direct}
\alias{nma_direct}
\title{Extract direct evidence when supported by the backend}
\usage{
nma_direct(fit, ...)
}
\arguments{
\item{fit}{An \code{nmaflow_fit} object.}

\item{...}{Backend arguments.}
}
\description{
Extract direct evidence when supported by the backend
}
