% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{nma_direction}
\alias{nma_direction}
\title{Harmonize effect direction}
\usage{
nma_direction(x, higher_is_better = TRUE, target = c("higher", "lower"))
}
\arguments{
\item{x}{Numeric outcome or effect vector.}

\item{higher_is_better}{Logical indicating the original favorable direction.}

\item{target}{Either \code{higher} or \code{lower}.}
}
\value{
Direction-harmonized numeric vector.
}
\description{
Harmonize effect direction
}
