% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-nonlinear.R
\name{nma_dose_compare}
\alias{nma_dose_compare}
\title{Compare dose-response candidate models}
\usage{
nma_dose_compare(fits)
}
\arguments{
\item{fits}{Named list of dose-response fits.}
}
\value{
Data frame of available information criteria.
}
\description{
Compare dose-response candidate models
}
