% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-nonlinear.R
\name{nma_dose_curve}
\alias{nma_dose_curve}
\title{Predict dose-response curves}
\usage{
nma_dose_curve(fit, ...)
}
\arguments{
\item{fit}{Dose-response \code{nmaflow_fit}.}

\item{...}{Passed to the backend prediction method.}
}
\description{
Predict dose-response curves
}
