% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-regression.R
\name{nma_effect_modifier}
\alias{nma_effect_modifier}
\title{Summarize candidate effect modifiers before meta-regression}
\usage{
nma_effect_modifier(x, modifiers)
}
\arguments{
\item{x}{An \code{nmaflow_data} object.}

\item{modifiers}{Candidate effect modifiers.}
}
\value{
The transitivity summary plus missingness by modifier.
}
\description{
Summarize candidate effect modifiers before meta-regression
}
