% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{nma_export}
\alias{nma_export}
\title{Export data and analysis templates to external ecosystems}
\usage{
nma_export(
  x,
  engine = c("python", "julia", "sas", "stata", "mplus", "jags", "bugs"),
  dir = "nma_export",
  prefix = "nmaflow"
)
}
\arguments{
\item{x}{\code{nmaflow_data}, \code{nmaflow_effects}, or data frame.}

\item{engine}{\code{python}, \code{julia}, \code{sas}, \code{stata}, \code{mplus}, \code{jags}, or \code{bugs}.}

\item{dir}{Output directory.}

\item{prefix}{File prefix.}
}
\value{
Paths created. Templates are starting points and must be scientifically reviewed.
}
\description{
Export data and analysis templates to external ecosystems
}
