% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{nma_fit_additive}
\alias{nma_fit_additive}
\title{Fit additive component NMA using netmeta}
\usage{
nma_fit_additive(fit, inactive = NULL, sep_components = "+", ...)
}
\arguments{
\item{fit}{A standard netmeta-backed \code{nmaflow_fit}.}

\item{inactive}{Optional inactive component.}

\item{sep_components}{Component separator.}

\item{...}{Passed to \code{netmeta::netcomb()}.}
}
\value{
An \code{nmaflow_fit} component model.
}
\description{
Fit additive component NMA using netmeta
}
