% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{nma_fit_freq}
\alias{nma_fit_freq}
\title{Fit frequentist network meta-analysis with netmeta}
\usage{
nma_fit_freq(
  x,
  studlab = "study",
  treat1 = "treat1",
  treat2 = "treat2",
  TE = "TE",
  seTE = "seTE",
  sm = NULL,
  common = FALSE,
  random = TRUE,
  reference = NULL,
  method_tau = "REML",
  ...
)
}
\arguments{
\item{x}{Contrast data from \code{\link[=nma_effects]{nma_effects()}} or a data frame.}

\item{studlab, treat1, treat2, TE, seTE}{Column names.}

\item{sm}{Summary measure passed to \code{netmeta}.}

\item{common, random}{Fit common- and/or random-effects models.}

\item{reference}{Reference treatment.}

\item{method_tau}{Heterogeneity estimator.}

\item{...}{Additional arguments passed to \code{netmeta::netmeta()}.}
}
\value{
An \code{nmaflow_fit} object.
}
\description{
Fit frequentist network meta-analysis with netmeta
}
