% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{nma_freq_summary}
\alias{nma_freq_summary}
\title{Frequentist model summary}
\usage{
nma_freq_summary(fit, ...)
}
\arguments{
\item{fit}{An \code{nmaflow_fit} object.}

\item{...}{Passed to \code{summary()}.}
}
\description{
Frequentist model summary
}
