% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{nma_league}
\alias{nma_league}
\title{Generate a league table}
\usage{
nma_league(fit, ...)
}
\arguments{
\item{fit}{An \code{nmaflow_fit}.}

\item{...}{Backend arguments.}
}
\description{
Generate a league table
}
