% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex-evidence.R
\name{nma_mlnmr}
\alias{nma_mlnmr}
\title{Combine IPD and aggregate evidence for ML-NMR preparation}
\usage{
nma_mlnmr(ipd_network, agd_network, reference = NULL, integration = NULL)
}
\arguments{
\item{ipd_network}{A network created by \code{\link[=nma_ipd]{nma_ipd()}} or \code{multinma::set_ipd()}.}

\item{agd_network}{Aggregate network created by \code{\link[=nma_make_multinma]{nma_make_multinma()}} or a multinma set function.}

\item{reference}{Optional common reference treatment.}

\item{integration}{Optional function that accepts the combined network and returns a
network with integration points added by \code{multinma::add_integration()}.}
}
\value{
Combined multinma network, optionally with integration points.
}
\description{
Combine IPD and aggregate evidence for ML-NMR preparation
}
