% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{nma_multiarm_check}
\alias{nma_multiarm_check}
\title{Check multi-arm evidence structure}
\usage{
nma_multiarm_check(x)
}
\arguments{
\item{x}{An \code{nmaflow_data} object.}
}
\value{
Data frame with arm counts and flags by study.
}
\description{
Check multi-arm evidence structure
}
