% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nma_nodesplit}
\alias{nma_nodesplit}
\title{Node-splitting convenience wrapper}
\usage{
nma_nodesplit(fit, ...)
}
\arguments{
\item{fit}{An \code{nmaflow_fit}.}

\item{...}{Backend arguments.}
}
\description{
Node-splitting convenience wrapper
}
