% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{nma_nodesplit_bayes}
\alias{nma_nodesplit_bayes}
\title{Fit Bayesian node-splitting models}
\usage{
nma_nodesplit_bayes(network, ...)
}
\arguments{
\item{network}{A multinma network.}

\item{...}{Passed to \code{\link[=nma_fit_bayes]{nma_fit_bayes()}}.}
}
\description{
Fit Bayesian node-splitting models
}
