% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-nonlinear.R
\name{nma_nonlinear}
\alias{nma_nonlinear}
\title{Create or fit agronomically useful nonlinear dose-response specifications}
\usage{
nma_nonlinear(
  data,
  model = c("mitscherlich", "quadratic", "emax", "custom"),
  run = TRUE,
  custom_formula = NULL,
  ...
)
}
\arguments{
\item{data}{Arm-level MBNMAdose data.}

\item{model}{\code{mitscherlich}, \code{quadratic}, \code{emax}, or \code{custom}.}

\item{run}{If \code{FALSE}, return only the model specification. If \code{TRUE}, fit with MBNMAdose.}

\item{custom_formula}{Custom MBNMAdose dose-response formula when \code{model='custom'}.}

\item{...}{Passed to \code{\link[=nma_dose]{nma_dose()}}.}
}
\value{
Specification list or fitted model.
}
\description{
Create or fit agronomically useful nonlinear dose-response specifications
}
