% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{nma_outliers}
\alias{nma_outliers}
\title{Outlier-detection orchestration}
\usage{
nma_outliers(fit, engine = c("influence", "NMAoutlier"), ...)
}
\arguments{
\item{fit}{A fitted model.}

\item{engine}{\code{NMAoutlier} or \code{influence}.}

\item{...}{Backend arguments.}
}
\description{
Outlier-detection orchestration
}
