% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{nma_pp_check}
\alias{nma_pp_check}
\title{Posterior/model-fit checking for multinma models}
\usage{
nma_pp_check(fit, type = c("dic", "loo", "waic", "summary"), ...)
}
\arguments{
\item{fit}{Bayesian \code{nmaflow_fit}.}

\item{type}{One of \code{dic}, \code{loo}, \code{waic}, or \code{summary}.}

\item{...}{Backend arguments.}
}
\description{
Posterior/model-fit checking for multinma models
}
