% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{nma_predict}
\alias{nma_predict}
\title{Predict treatment effects or intervals using the fitted backend}
\usage{
nma_predict(fit, ...)
}
\arguments{
\item{fit}{An \code{nmaflow_fit} object.}

\item{...}{Passed to the backend prediction method.}
}
\description{
Predict treatment effects or intervals using the fitted backend
}
