% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-architecture.R
\name{nma_preflight}
\alias{nma_preflight}
\title{Preflight check for NMA readiness}
\usage{
nma_preflight(x, ...)
}
\arguments{
\item{x}{An \code{nmaflow_data} object.}

\item{...}{Passed to \code{\link[=nma_validate]{nma_validate()}}.}
}
\value{
An object of class \code{nmaflow_preflight} with GREEN/YELLOW/RED status.
}
\description{
Preflight check for NMA readiness
}
