% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{nma_prior}
\alias{nma_prior}
\title{Construct a prior distribution for multinma}
\usage{
nma_prior(
  distribution = c("normal", "half_normal", "student_t", "half_student_t", "cauchy",
    "half_cauchy", "exponential"),
  ...
)
}
\arguments{
\item{distribution}{One of \code{normal}, \code{half_normal}, \code{student_t}, \code{half_student_t}, \code{cauchy}, \code{half_cauchy}, or \code{exponential}.}

\item{...}{Parameters passed to the corresponding multinma prior constructor.}
}
\value{
A multinma prior object.
}
\description{
Construct a prior distribution for multinma
}
