% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{nma_rank}
\alias{nma_rank}
\title{Rank treatments while retaining uncertainty information}
\usage{
nma_rank(fit, small_values = "undesirable", probabilities = TRUE, ...)
}
\arguments{
\item{fit}{An \code{nmaflow_fit}.}

\item{small_values}{Direction setting passed to netmeta.}

\item{probabilities}{For multinma, return posterior rank probabilities instead of ranks.}

\item{...}{Backend arguments.}
}
\description{
Rank treatments while retaining uncertainty information
}
