% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{nma_rank_uncertainty}
\alias{nma_rank_uncertainty}
\title{Summarize uncertainty in treatment ranking}
\usage{
nma_rank_uncertainty(x)
}
\arguments{
\item{x}{Result from \code{\link[=nma_rank]{nma_rank()}} or \code{\link[=nma_boot_rank]{nma_boot_rank()}}.}
}
\value{
A backend-neutral summary where feasible.
}
\description{
Summarize uncertainty in treatment ranking
}
