% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-regression.R
\name{nma_regression_compare}
\alias{nma_regression_compare}
\title{Compare a set of network meta-regression models}
\usage{
nma_regression_compare(fits)
}
\arguments{
\item{fits}{Named list of backend models or nmaFlow fits.}
}
\value{
Data frame of AIC/BIC when available.
}
\description{
Compare a set of network meta-regression models
}
