% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{nma_reproduce}
\alias{nma_reproduce}
\title{Save a reproducibility bundle}
\usage{
nma_reproduce(x, dir = "nmaflow_repro", include_raw = TRUE)
}
\arguments{
\item{x}{nmaFlow object.}

\item{dir}{Bundle directory.}

\item{include_raw}{Include raw imported data when present.}
}
\value{
Paths created.
}
\description{
Save a reproducibility bundle
}
