% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{nma_session}
\alias{nma_session}
\title{Reproducibility session report}
\usage{
nma_session()
}
\value{
A list containing R session, nmaFlow options and optional backend versions.
}
\description{
Reproducibility session report
}
