% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{nma_threshold}
\alias{nma_threshold}
\title{Decision-threshold analysis wrapper}
\usage{
nma_threshold(fit, ...)
}
\arguments{
\item{fit}{Fitted NMA model.}

\item{...}{Passed to an available \code{nmathresh} entry function.}
}
\description{
Decision-threshold analysis wrapper
}
