% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nma_transitivity}
\alias{nma_transitivity}
\title{Assess transitivity using effect-modifier distributions}
\usage{
nma_transitivity(x, modifiers, by = c("treatment", "coarse", "fine"))
}
\arguments{
\item{x}{An \code{nmaflow_data} object.}

\item{modifiers}{Character vector of effect-modifier columns.}

\item{by}{Treatment node representation.}
}
\value{
An object with treatment-specific summaries and overlap flags.
}
\description{
Assess transitivity using effect-modifier distributions
}
