% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmaFlow-package.R, R/zzz.R
\docType{package}
\name{nmaFlow-package}
\alias{nmaFlow-package}
\title{nmaFlow: Integrated Workflows for Network Meta-Analysis}
\description{
\code{nmaFlow} is a design-aware orchestration and teaching layer for network
meta-analysis. It emphasizes auditable data architecture, conservative network
formation, frequentist and Bayesian backends, study-level bootstrap, transitivity
and inconsistency diagnostics, advanced evidence structures, agronomic simulation,
publication-oriented outputs, and reproducibility.

Integrated workflows for Network Meta-Analysis.
}
\author{
\strong{Maintainer}: Walter Esfrain Pereira \email{walter@cca.ufpb.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Walter Esfrain Pereira \email{walter@cca.ufpb.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]
  \item Magali Haidee Pereira Martinez (\href{https://orcid.org/0009-0009-5419-959X}{ORCID})
}

}
\keyword{internal}
